feat: move all programs to program folder and move porgram's account to state file
