+++ program/token/instruction.go
