+++ program/token/state_test.go
