--- tokenprog/multisig_account.go
@@ -1,3 +1,0 @@
-package tokenprog
-
-const MultisigAccountSize uint64 = 355
